requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
-e git+https://github.com/deimos-deimos/comfy_api_simplified.git@302e9d814b93449da811f179b54bba25960b75ef#egg=comfy_api_simplified
idna==3.6
pillow==10.2.0
rel==0.4.9.6
requests==2.31.0
urllib3==2.2.0
websocket-client==1.7.0
